Move critic agent loop from generic location to critic/
The agent loop in props/core/agent_loop/ was critic-specific,
importing critic tools. Moved to props/core/critic/loop.py
where it belongs.
