--- props/core/agent_loop/BUILD.bazel
@@ -1,21 +1,0 @@
-"""In-container agent loop for props agents."""
-
-load("@rules_python//python:defs.bzl", "py_library")
-
-py_library(
-    name = "loop",
-    srcs = ["loop.py"],
-    imports = ["../../.."],
-    visibility = ["//props:__subpackages__"],
-    deps = [
-        "//agent_core",
-        "//agent_core:direct_provider",
-        "//mcp_infra/exec:subprocess_exec",
-        "//openai_utils",
-        "//props/core:agent_helpers",
-        "//props/core/critic:tools",
-        "//props/core/db:session",
-        "@pypi//openai",
-        "@pypi//pydantic",
-    ],
-)
